Multiple corrections to evaluations

+++ MMOEASA/mmoeasaSolution.py
@@ -10,7 +10,7 @@ class MMOEASASolution(Solution):
     def __init__(self, _id: int=None, vehicles: List[Vehicle]=None, feasible: bool=True, default_temperature: float=0.0, temperature: float=0.0, cooling_rate: float=0.0, total_distance: float=0.0, distance_unbalance: float=0.0, cargo_unbalance: float=0.0, rank: int=INT_MAX) -> None:
         super(MMOEASASolution, self).__init__(_id=_id, vehicles=vehicles, feasible=feasible, default_temperature=default_temperature, temperature=temperature, cooling_rate=cooling_rate, total_distance=total_distance, rank=rank)
         self.distance_unbalance: float=float(distance_unbalance)
-        self.cargo_unbalance: float=float(cargo_unbalance)
+        self.cargo_unbalance: int=int(cargo_unbalance)
 
     def __str__(self) -> str:
         return f"total_distance={self.total_distance}, distance_unbalance={self.distance_unbalance}, cargo_unbalance={self.cargo_unbalance}, {len(self.vehicles)=}, {[f'{i}. {str(v)}' for i, v in enumerate(self.vehicles)]}"
@@ -19,7 +19,7 @@ class MMOEASASolution(Solution):
         self.feasible = False
         self.total_distance = float(INFINITY)
         self.distance_unbalance = float(INFINITY)
-        self.cargo_unbalance = float(INFINITY)
+        self.cargo_unbalance = INFINITY
 
     def objective_function(self, instance: ProblemInstance) -> None:
         self.total_distance = 0.0
@@ -38,8 +38,8 @@ class MMOEASASolution(Solution):
 
         minimum_distance = float(INFINITY)
         maximum_distance = 0.0
-        minimum_cargo = float(INFINITY)
-        maximum_cargo = 0.0
+        minimum_cargo = INFINITY
+        maximum_cargo = 0
         for vehicle in self.vehicles:
             # these cannot be converted to "if ... elif" because we may miss, for example, our "maximum_distance" as on the first iteration it will also be less than "INFINITY" ("minimum_distance")
             if vehicle.route_distance < minimum_distance:
